Bug 1493833 - Fix ESLint errors with import/no-cycle
