+++ tests/ui/unit/models/jobs.tests.js
@@ -1,7 +1,7 @@
 import * as fetchMock from 'fetch-mock';
 
 import JobModel from '../../../../ui/models/job';
-import { getProjectUrl } from '../../../../ui/helpers/url';
+import { getProjectUrl } from '../../../../ui/helpers/location';
 import jobListFixtureOne from '../../mock/job_list/job_1';
 import paginatedJobListFixtureOne from '../../mock/job_list/pagination/page_1';
 import paginatedJobListFixtureTwo from '../../mock/job_list/pagination/page_2';
