+++ ui/models/textLogStep.js
@@ -1,4 +1,4 @@
-import { getProjectJobUrl } from '../helpers/url';
+import { getProjectJobUrl } from '../helpers/location';
 
 export default class TextLogStepModel {
   static get(jobId) {
